fix(enumerate): prioritize alphanumeric chars
Previously, `enumerate` would output characters in the "unicode order".
That means, characters which are early in the unicode range are output
first (such as "\u0000", "\u0001"). Now, the more common alphanumeric
characters are always produced first. That way, we get the more
representative strings tend to get enumerated first, which is useful
because we usually only look at the first few items in the enumeration.